+++ src/filesystem/stream.hpp
@@ -16,7 +16,7 @@ namespace arda
 		inline int getPosition() { return m_position; }
 
 		virtual void seek(int offset, SeekOrigin origin) = 0;
-		virtual int read(char* buffer,int numBytes) = 0;
+		virtual unsigned int read(char* buffer,size_t numBytes) = 0;
 	protected:
 		int m_position;
 		int m_size;
